+++ init.php
@@ -95,7 +95,10 @@ require __DIR__ . '/lib/BaseStripeClient.php';
 require __DIR__ . '/lib/StripeClient.php';
 
 // EventRouter
+require __DIR__ . '/lib/UnhandledNotificationDetails.php';
+require __DIR__ . '/lib/AbstractEventNotificationHandler.php';
 require __DIR__ . '/lib/StripeEventNotificationHandler.php';
+require __DIR__ . '/lib/StripeEventNotificationHandlerWithoutVerification.php';
 
 // The beginning of the section generated from our OpenAPI spec
 require __DIR__ . '/lib/Account.php';
