+++ lib/UnhandledNotificationDetails.php
@@ -1,0 +1,16 @@
+<?php
+
+// File copied from our code generator; changes here will be overwritten.
+
+namespace Stripe;
+
+class UnhandledNotificationDetails
+{
+    /** @var bool whether the SDK has types for this event */
+    public $isKnownEventType;
+
+    public function __construct($isKnownEventType)
+    {
+        $this->isKnownEventType = $isKnownEventType;
+    }
+}
